change data folder to constants, move menu data to constants folder, clean AddBookForm
