+++ src/navigation/DrawerNavigator.tsx
@@ -2,7 +2,7 @@ import {createDrawerNavigator} from '@react-navigation/drawer';
 import {Bookshelf} from '@src/screens/Bookshelf';
 import {Home} from '@src/screens/Home';
 import {DrawerCustomerView} from './DrawerCustomerView';
-import {SCREENS} from '@src/data/screenEnums';
+import {SCREENS} from '@src/constants/screenEnums';
 import {styles} from './styles';
 
 const {Navigator, Screen} = createDrawerNavigator();
